+++ ReleaseNotes.md
@@ -1,3 +1,5 @@
 # Release Notes
 
-## 1.2.0-beta
+## 1.2.0
+* symbol packages retain the symbols extension when updating the name
+* Added nuspec frameworkassemblies add command